consolidate spheregroup code
