--- pydl/pydlutils/spheregroup/__init__.py
@@ -1,9 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This subpackage corresponds to the spheregroup directory of idlutils.
-"""
-from .chunks import chunks
-from .groups import groups
-from .spheregroup import spheregroup
-from .spherematch import spherematch
