+++ setup.cfg
@@ -13,7 +13,7 @@ norecursedirs = build docs/_build
 doctest_plus = enabled
 
 [pep8]
-ignore = E12,E226
+ignore = E12,E226,E501
 
 [ah_bootstrap]
 auto_use = True
